1 eventlistener for DOM now
